timesheet data validation and security

+++ src/main/frontend/src/app/helpers/employeePortal.interceptor.ts
@@ -1,18 +1,23 @@
 import { HttpErrorResponse, HttpEvent, HttpHandler, HttpHeaders, HttpInterceptor, HttpRequest } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable, of } from 'rxjs';
-import { catchError, tap } from 'rxjs/operators';
+import { catchError, first, tap } from 'rxjs/operators';
 import { AuthenticationService } from '../services/authentication.service';
 import { Router } from '@angular/router';
 import { LoaderService } from '../services/loader.service';
+import { User } from '../models/user';
 
 @Injectable()
 export class EmployeePortalInterceptor implements HttpInterceptor {
-
+currentUser:User = new User();
     constructor(
         private authenticationService : AuthenticationService,private loaderService: LoaderService,
         private router: Router
-    ) { }
+    ) { 
+        this.authenticationService.currentUser.subscribe(x => {
+      this.currentUser = x;
+    });
+    }
 
     intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
         const token: string = sessionStorage.getItem('token');
@@ -48,18 +53,39 @@ export class EmployeePortalInterceptor implements HttpInterceptor {
     }
     
     userLogout(){
-        this.authenticationService.stopUserSessionCheck();
-          sessionStorage.removeItem('currentUser');
-          sessionStorage.removeItem('token');
-          sessionStorage.removeItem('logInfo');
-          sessionStorage.removeItem('maxFileSize');
-          sessionStorage.removeItem('maxRequestSize');
-          sessionStorage.removeItem('sessioncheck');
-          sessionStorage.removeItem('breadcrumb');
-          // delete method call for cookies
-          this.authenticationService.deleteCookies();
-          this.authenticationService.setcurrentUserSubject(null);
-          this.router.navigate(['/login']);
-          setTimeout(() => {location.reload();});   
+        let user = new User();
+        user.empId = this.currentUser.empId;
+            this.authenticationService.logoutUser(user).pipe(first()).subscribe((response: any) => {
+              if (response.serviceStatus == "Success") {
+                this.authenticationService.stopUserSessionCheck();
+                //console.log(response.serviceResponse);
+                  sessionStorage.removeItem('currentUser');
+                  sessionStorage.removeItem('token');
+                  sessionStorage.removeItem('logInfo');
+                  sessionStorage.removeItem('maxFileSize');
+                  sessionStorage.removeItem('maxRequestSize');
+                  sessionStorage.removeItem('sessioncheck');
+                  sessionStorage.removeItem('breadcrumb');
+                // delete method call for cookies
+                this.authenticationService.deleteCookies();
+                this.authenticationService.setcurrentUserSubject(null);
+                this.router.navigate(['/login']);
+                setTimeout(() => {location.reload();});
+              }
+        },);
+
+        // this.authenticationService.stopUserSessionCheck();
+        //   sessionStorage.removeItem('currentUser');
+        //   sessionStorage.removeItem('token');
+        //   sessionStorage.removeItem('logInfo');
+        //   sessionStorage.removeItem('maxFileSize');
+        //   sessionStorage.removeItem('maxRequestSize');
+        //   sessionStorage.removeItem('sessioncheck');
+        //   sessionStorage.removeItem('breadcrumb');
+        //   // delete method call for cookies
+        //   this.authenticationService.deleteCookies();
+        //   this.authenticationService.setcurrentUserSubject(null);
+        //   this.router.navigate(['/login']);
+        //   setTimeout(() => {location.reload();});   
       }
 }
